requirements.txt
six==1.10.0
websocket-client==0.35.0
wheel==0.24.0
